chore: making SSM code minimal
A lot of the ssm.py code was very experimental on my end... there is also very little consensus. Only keeping core SSM Block logic.
