+++ examples/shakespeare.py
@@ -2,7 +2,7 @@
 """Trains a state space model on a character-level tokenized dataset of Shakespeare."""
 
 from dataclasses import dataclass
-from typing import Any, Iterator, Protocol
+from typing import Any, Iterator, Literal, Protocol
 
 import equinox as eqx
 import jax
@@ -12,6 +12,7 @@ import tensorflow_datasets as tfds
 from jaxtyping import Array, PRNGKeyArray, PyTree
 
 import xax
+from xax.nn.ssm import BaseSSMBlock, DiagSSMBlock, SSMBlock
 
 
 @dataclass(frozen=True)
@@ -200,6 +201,97 @@ class LSTM(eqx.Module):
 
         return sequence
 
+class S4(eqx.Module):
+    vocab_embedding: eqx.nn.Embedding
+    proj_in: eqx.nn.Linear
+    proj_out: eqx.nn.Linear
+    blocks: list[BaseSSMBlock]
+    num_layers: int = eqx.static_field()
+    hidden_size: int = eqx.static_field()
+    skip_connections: bool = eqx.static_field()
+
+    def __init__(
+        self,
+        input_size: int,
+        hidden_size: int,
+        output_size: int,
+        num_layers: int,
+        block_type: Literal["ssm", "diag"] = "ssm",
+        skip_connections: bool = False,
+        *,
+        key: PRNGKeyArray,
+    ) -> None:
+        vocab_key, s4_key = jax.random.split(key, 2)
+        self.vocab_embedding = eqx.nn.Embedding(input_size, hidden_size, key=vocab_key)
+        self.proj_in = eqx.nn.Linear(hidden_size, hidden_size, key=key)
+        self.proj_out = eqx.nn.Linear(hidden_size, output_size, key=key)
+
+        block_keys = jax.random.split(s4_key, num_layers)
+
+        def get_block(key: PRNGKeyArray) -> BaseSSMBlock:
+            match block_type:
+                case "ssm":
+                    return SSMBlock(hidden_size, key=key)
+                case "diag":
+                    return DiagSSMBlock(hidden_size, key=key)
+                case _:
+                    raise ValueError(f"Unknown block type: {block_type}")
+
+        self.blocks = [get_block(block_keys[i]) for i in range(num_layers)]
+        self.skip_connections = skip_connections
+        self.num_layers = num_layers
+        self.hidden_size = hidden_size
+
+    def __call__(self, hs: list[Array], x: Array) -> tuple[list[Array], Array]:
+        new_hs = []
+        for i, block in enumerate(self.blocks):
+            h = block.forward(hs[i], x)
+            new_hs.append(h)
+            xh = jax.nn.gelu(h)
+            x = xh + x if self.skip_connections else xh
+        y = self.proj_out(x)
+        return new_hs, y
+
+    def _embed_input(self, x: Array) -> Array:
+        """U is the input to the S4 cell."""
+        embedded = self.vocab_embedding(x)
+        return jax.nn.gelu(self.proj_in(embedded))
+
+    def predict_sequence(self, x_seq: Array) -> Array:
+        x_emb = jax.vmap(self._embed_input)(x_seq)
+        hs = [jnp.zeros(self.hidden_size) for _ in range(self.num_layers)]
+
+        def step(hs: list[Array], x: Array) -> tuple[list[Array], Array]:
+            hs, y = self(hs, x)
+            return hs, y
+
+        _, y_seq = jax.lax.scan(step, hs, x_emb)
+        return y_seq
+
+    def generate_sequence(self, prompt_seq: Array, max_len: int) -> Array:
+        hs = [jnp.zeros(self.hidden_size) for _ in range(self.num_layers)]
+        prompt_seq_embedded = jax.vmap(self._embed_input)(prompt_seq)
+
+        def encode_step(hs: list[Array], x: Array) -> tuple[list[Array], Array]:
+            hs, y = self(hs, x)
+            return hs, y
+
+        def decode_step(
+            carry: tuple[list[Array], Array, PRNGKeyArray],
+            _: Any,
+        ) -> tuple[tuple[list[Array], Array, PRNGKeyArray], Array]:
+            hs, last_token, rng = carry
+            token_embedded = self._embed_input(last_token)
+            hs, y = self(hs, token_embedded)
+            token = jax.random.categorical(rng, y)
+            rng = jax.random.split(rng)[0]
+            return (hs, token, rng), token
+
+        hs, _ = jax.lax.scan(encode_step, hs, prompt_seq_embedded)
+        _, sequence = jax.lax.scan(decode_step, (hs, prompt_seq[-1], jax.random.PRNGKey(0)), None, length=max_len)
+
+        return sequence
+
 
 class ShakespearePrediction(xax.Task[Config]):
     def __init__(self, config: Config) -> None:
@@ -242,7 +334,7 @@ class ShakespearePrediction(xax.Task[Config]):
                     key=key,
                 )
             case "s4":
-                return xax.S4(
+                return S4(
                     input_size=self.config.input_size,
                     hidden_size=self.config.hidden_size,
                     output_size=self.config.output_size,
@@ -326,5 +418,5 @@ class ShakespearePrediction(xax.Task[Config]):
 if __name__ == "__main__":
     # Launch the training task.
     ShakespearePrediction.launch(Config(
-        model_type="lstm",
+        model_type="s4",
     ))
